[backend] feat(threat-arsenal): import threat arsenal
