+++ openaev-api/src/main/java/io/openaev/api/payload/PayloadApiImporter.java
@@ -1,14 +1,20 @@
 package io.openaev.api.payload;
 
-import static java.util.Collections.emptyList;
+import static io.openaev.helper.StreamHelper.fromIterable;
+import static io.openaev.helper.StreamHelper.iterableToSet;
 
 import io.openaev.aop.AccessControl;
 import io.openaev.database.model.Action;
 import io.openaev.database.model.Payload;
 import io.openaev.database.model.ResourceType;
+import io.openaev.database.repository.AttackPatternRepository;
+import io.openaev.database.repository.DomainRepository;
+import io.openaev.database.repository.TagRepository;
 import io.openaev.jsonapi.IncludeOptions;
 import io.openaev.jsonapi.IncludeOptions.IncludeMode;
 import io.openaev.jsonapi.JsonApiDocument;
+import io.openaev.jsonapi.Relationship;
+import io.openaev.jsonapi.ResourceIdentifier;
 import io.openaev.jsonapi.ResourceObject;
 import io.openaev.jsonapi.ZipJsonApi;
 import io.openaev.rest.helper.RestBehavior;
@@ -18,8 +24,9 @@ import io.openaev.service.ImportService;
 import io.openaev.service.ZipJsonService;
 import io.swagger.v3.oas.annotations.Operation;
 import jakarta.validation.constraints.NotNull;
+import java.util.Collections;
+import java.util.List;
 import java.util.Map;
-import java.util.Set;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.MediaType;
@@ -49,47 +56,64 @@ public class PayloadApiImporter extends RestBehavior {
   private final ZipJsonApi<Payload> zipJsonApi;
   private final ImportService importService;
   private final PayloadService payloadService;
+  private final AttackPatternRepository attackPatternRepository;
+  private final DomainRepository domainRepository;
+  private final TagRepository tagRepository;
 
+  /**
+   * Imports a payload from a JSON:API document (legacy format).
+   *
+   * <p>Legacy payload exports may contain {@code payload_attack_patterns}, {@code payload_domains},
+   * and {@code payload_tags} as relationships. Since these fields now live on {@code
+   * InjectorContract}, they are extracted from the document and passed to the synchronisation
+   * method that creates/updates the associated injector contract.
+   */
   @Operation(
       description =
-          "Imports a payload from a JSON:API document. The name will be suffixed with '(import)' by default.")
+          "Imports a payload from a JSON:API document. The name will be suffixed with '(Import)' by default.")
   @PostMapping(
       value = "/import",
       consumes = MediaType.MULTIPART_FORM_DATA_VALUE,
       produces = MediaType.APPLICATION_JSON_VALUE)
-  @Transactional
+  @Transactional(rollbackFor = Exception.class)
   @AccessControl(actionPerformed = Action.WRITE, resourceType = ResourceType.PAYLOAD)
   public ResponseEntity<JsonApiDocument<ResourceObject>> importJson(
       @RequestPart("file") @NotNull MultipartFile file) throws Exception {
     try {
-      ZipJsonService.ImportOutput<Payload> response =
-          zipJsonApi.handleImport(file, "payload_name", IMPORT_OPTIONS, this::sanitize);
+      ZipJsonService.ImportOutput<Payload> response = zipJsonApi.handleImport(file, "payload_name", IMPORT_OPTIONS, null);
+
+      // Extract legacy relationship IDs from the original source document.
+      // These fields (attack_patterns, domains, tags) no longer exist on Payload
+      // but may be present in legacy exports — they now live on InjectorContract.
+      Map<String, Relationship> rels =
+          response.sourceDocument().data().relationships() != null
+              ? response.sourceDocument().data().relationships()
+              : Collections.emptyMap();
+
+      List<String> attackPatternIds = extractRelationshipIds(rels, "payload_attack_patterns");
+      List<String> domainIds = extractRelationshipIds(rels, "payload_domains");
+      List<String> tagIds = extractRelationshipIds(rels, "payload_tags");
+
       payloadService.synchroniseInjectorContractBasedOnPayload(
-          response.persistedData(), emptyList(), Set.of(), Set.of());
+          response.persistedData(),
+          fromIterable(attackPatternRepository.findAllById(attackPatternIds)),
+          iterableToSet(domainRepository.findAllById(domainIds)),
+          iterableToSet(tagRepository.findAllById(tagIds)));
+
       return ResponseEntity.ok(response.jsonApiDocument());
     } catch (Exception ex) {
       log.warn("Fallback to old import due to {}", ex.getMessage(), ex);
+      // Fall back to the legacy importer
       importService.handleFileImport(file, null, null);
       return ResponseEntity.ok().build();
     }
   }
 
-  /**
-   * Removes detection remediations whose collector does not exist in the target database. During
-   * import, the collector is resolved by its business key (type). If not found, the collector field
-   * is null and the remediation must be dropped to avoid constraint violations.
-   */
-  private Payload sanitize(Payload payload) {
-    payload
-        .getDetectionRemediations()
-        .removeIf(
-            dr -> {
-              if (dr.getCollectorType() == null) {
-                log.warn("Skipping detection remediation — collector not found in target database");
-                return true;
-              }
-              return false;
-            });
-    return payload;
+  private List<String> extractRelationshipIds(Map<String, Relationship> rels, String relName) {
+    Relationship rel = rels.get(relName);
+    if (rel == null) {
+      return Collections.emptyList();
+    }
+    return rel.asMany().stream().map(ResourceIdentifier::id).toList();
   }
 }
