+++ openaev-api/src/main/java/io/openaev/api/threat_arsenal/dto/ThreatArsenalAction.java
@@ -1,4 +1,4 @@
-package io.openaev.rest.threat_arsenal.dto;
+package io.openaev.api.threat_arsenal.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 import io.openaev.database.model.Endpoint;
