+++ openaev-api/src/main/java/io/openaev/api/threat_arsenal/dto/ThreatArsenalActionWithContentOutput.java
