+++ openaev-front/src/utils/lang/fr.json
@@ -998,7 +998,7 @@
   "I have read and comply with the above statement": "J'ai lu et j'accepte la déclaration ci-dessus",
   "If you remove the default email address, the email reception for this simulation / scenario will be disabled.": "Si vous supprimez l'adresse e-mail par défaut, la réception des e-mails pour cette simulation / ce scénario sera désactivée.",
   "IMAP": "IMAP",
-  "Import": "Importation",
+  "Import": "Importer",
   "Import a custom dashboard": "Importer un tableau de bord personnalisé",
   "Import a simulation": "Importer une simulation",
   "Import endpoints": "Importer des endpoints",
